requirements.txt
bottle==0.12.13
future==0.16.0
oauthlib==2.0.6
python-twitter==3.3
requests==2.18.4
requests-oauthlib==0.8.0